Added spawners in the mainscene.
GameStateMachine - Added references to player one and twos spawner. added a function that will find the middle of each players side and set the player spawnpoint on spawner. added a function that turn on the spawners.
PlayState - Added function: The spawners are turned on through gameStateMachine
MainScene - in the inspector the gamestatemachine no gets references to the spawners in the inspector